chore: reorganize backend folder structure — Phase 5 EE mirror (methods)
Renames ee/server/methods → ee/server/meteor-methods and moves EE method
files there: canned-responses methods (still registered via the
License.onLicense dynamic imports), license methods.ts (as license.ts) and
livechat-enterprise removeBusinessHour into omnichannel/. All EE sources
stay under ee/ per the license boundary.
