+++ apps/meteor/ee/app/canned-responses/server/index.ts
@@ -6,8 +6,8 @@ await License.onLicense('canned-responses', async () => {
 	await import('./hooks/onRemoveAgentDepartment');
 	await import('./hooks/onSaveAgentDepartment');
 	await import('./hooks/cannedResponses');
-	await import('./methods/saveCannedResponse');
-	await import('./methods/removeCannedResponse');
+	await import('../../../server/meteor-methods/saveCannedResponse');
+	await import('../../../server/meteor-methods/removeCannedResponse');
 
 	await createSettings();
 });
