+++ apps/meteor/ee/app/license/server/index.ts
@@ -1,2 +1,2 @@
 import './settings';
-import './methods';
+import '../../../server/meteor-methods/license';
