+++ apps/meteor/ee/server/meteor-methods/license.ts
@@ -4,7 +4,7 @@ import { License } from '@rocket.chat/license';
 import { check } from 'meteor/check';
 import { Meteor } from 'meteor/meteor';
 
-import { methodDeprecationLogger } from '../../../../app/lib/server/lib/deprecationWarningLogger';
+import { methodDeprecationLogger } from '../../../app/lib/server/lib/deprecationWarningLogger';
 
 declare module '@rocket.chat/ddp-client' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
